+++ source/NetworkStructure/NeuronLayer.h
@@ -11,9 +11,9 @@ class NeuronLayer
 public:
     NeuronLayer();
 
-    std::shared_ptr<Neuron> getNeuron(const unsigned int index) const;
+    std::shared_ptr<Neuron> getNeuron(const size_t index) const;
     std::vector<std::shared_ptr<Neuron>> getNeurons() const;
-    void setNeuron(const unsigned int index, std::shared_ptr<Neuron> neuron);
+    void setNeuron(const size_t index, std::shared_ptr<Neuron> neuron);
     void appendNeuron(std::shared_ptr<Neuron> neuron);
     void connectNextLayerNeurons(std::shared_ptr<NeuronLayer> nextLayer);
     void connectPrevLayerNeurons(std::shared_ptr<NeuronLayer> prevLayer);
